[refactor] #25 - 경유지 위치 설정 화면 뷰 파일 분리
