+++ Carmunication/Views/SessionList/GroupAdd/SelectPointMapView.swift
@@ -1,0 +1,48 @@
+//
+//  SelectPointMapView.swift
+//  Carmunication
+//
+//  Created by 허준혁 on 10/5/23.
+//
+
+import UIKit
+
+final class SelectPointMapView: UIView {
+
+    lazy var vStackContainer = {
+        let vStack = UIStackView(arrangedSubviews: [titleLabel, backButton])
+        vStack.axis = .vertical
+        vStack.alignment = .center
+        vStack.distribution = .fillEqually
+        return vStack
+    }()
+
+    var titleLabel = {
+        let label = UILabel()
+        label.text = "맵뷰"
+        return label
+    }()
+
+    var backButton = {
+        let backButton = UIButton()
+        backButton.setTitle("뒤로", for: .normal)
+        backButton.setBackgroundImage(.pixel(ofColor: .blue), for: .normal)
+        backButton.translatesAutoresizingMaskIntoConstraints = false
+        return backButton
+    }()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+    }
+
+    override func draw(_ rect: CGRect) {
+        addSubview(vStackContainer)
+        vStackContainer.snp.makeConstraints { make in
+            make.center.equalToSuperview()
+        }
+    }
+}
